chore: ♻️ moved code from Manager to Diagnostics
