+++ src/lib/components/Diagnostics/Demo.svelte
