+++ src/lib/components/Diagnostics/TerminateDB.svelte
